requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
docutils==0.20.1
idna==3.4
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
Pygments==2.15.1
requests==2.30.0
urllib3==2.0.2
